perfect number
A number is perfect if it is equal to the sum of its divisors, itself
excluded. For instance, the divisors
of 28 distinct from 28 are 1, 2, 4, 7 and 14, and 1 + 2 + 4 + 7 + 14 =
28, hence 28 is perfect.
Write a program that outputs all 3-digit perfect numbers. There is
actually a unique solution. The output of the program should be of the
form:
... is a solution.